backend/requirements.txt
black==22.10.0
coverage==6.5.0
Django==4.2.4
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.4
